% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical test.R
\name{ANOVA_Non_param}
\alias{ANOVA_Non_param}
\title{ANOVA_Non_param}
\usage{
ANOVA_Non_param(start, end, dataX, group, comparison)
}
\arguments{
\item{start}{the start of your matrix or data frame}

\item{end}{the end of the matrix or data frame}

\item{dataX}{data frame or matrix for the Kruskal Wallis test}

\item{group}{Group for the comparison}

\item{comparison}{name of your comparison that it will the title of your output}
}
\value{
the function give a csv file with p-value of the Kruskal Waliis test and the mean and SD for each group for all variable in your matrix or data frame
}
\description{
ANOVA_Non_param
}
\author{
Matthieu COQ
Version: 2.0
Date: 04-Jun-21
Objective: the Objective is to calculate the non-parametric ANOVA
}
