% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical test.R
\name{normality}
\alias{normality}
\title{@title  normality
 @author Matthieu COQ
Version: 2.0
Date: 04-Jun-21
Objective: the Objective is to test the normality of the data}
\usage{
normality(start, end, dataX, project)
}
\arguments{
\item{start}{the start of your matrix or data frame}

\item{end}{the end of the matrix or data frame}

\item{dataX}{data frame or matrix for the Shapiro-Wilk test}

\item{project}{name of your comparison that it will the title of your output}
}
\value{
the function give a csv file with p-value of the Shapiro-Wilk test for all variable in your matrix or data frame
}
\description{
@title  normality
 @author Matthieu COQ
Version: 2.0
Date: 04-Jun-21
Objective: the Objective is to test the normality of the data
}
