% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical test.R
\name{wilcox}
\alias{wilcox}
\title{wilcox}
\usage{
wilcox(start, end, group_1, group_2, log, base, comparison)
}
\arguments{
\item{start}{the start of your matrix or data frame}

\item{end}{the end of the matrix or data frame}

\item{group_1}{data frame or matrix for  the first group of your comparison for the Wilcoxon test}

\item{group_2}{data frame or matrix for  the second group of your comparison for the Wilcoxon test}

\item{log}{logical TRUE if your data are log transformed}

\item{base}{base for your log transformation}

\item{comparison}{name of your comparison that it will the title of your output}
}
\value{
the function give a csv file with a mean, the standard deviation, the p-value of a Wilcoxon test and the foldchange for all variable in your matrix or data frame
}
\description{
wilcox
}
\author{
Matthieu COQ
Version: 1.0
Date: 23-Aug-2018
Objective: the Objective is to calculate a Wilcoxon test and the foldchange
}
